New kas::layout::size_types private module
