+++ src/layout/mod.rs
@@ -38,6 +38,7 @@ mod grid_solver;
 mod row_solver;
 mod single_solver;
 mod size_rules;
+mod size_types;
 mod sizer;
 mod storage;
 
@@ -47,7 +48,8 @@ pub use align::{Align, AlignHints, CompleteAlignment};
 pub use grid_solver::{GridChildInfo, GridSetter, GridSolver};
 pub use row_solver::{RowPositionSolver, RowSetter, RowSolver};
 pub use single_solver::{SingleSetter, SingleSolver};
-pub use size_rules::{Margins, SizeRules, StretchPolicy};
+pub use size_rules::SizeRules;
+pub use size_types::{Margins, StretchPolicy};
 pub use sizer::{solve_size_rules, RulesSetter, RulesSolver, SolveCache};
 pub use storage::{
     DynGridStorage, DynRowStorage, FixedGridStorage, FixedRowStorage, GridStorage, RowStorage,
