+++ src/layout/size_rules.rs
@@ -9,6 +9,7 @@ use smallvec::SmallVec;
 use std::fmt;
 use std::iter::Sum;
 
+use super::{Margins, StretchPolicy};
 use crate::conv::{Cast, Conv};
 use crate::dir::Directional;
 use crate::geom::Size;
@@ -16,82 +17,6 @@ use crate::geom::Size;
 // for doc use
 #[allow(unused)]
 use kas::draw::SizeHandle;
-
-/// Margin sizes
-///
-/// Used by the layout system for margins around child widgets. Margins may be
-/// drawn in and handle events like any other widget area.
-#[derive(Copy, Clone, Debug, Default)]
-pub struct Margins {
-    /// Size of horizontal margins
-    pub horiz: (u16, u16),
-    /// Size of vertical margins
-    pub vert: (u16, u16),
-}
-
-impl Margins {
-    /// Zero-sized margins
-    pub const ZERO: Margins = Margins::splat(0);
-
-    /// Margins with equal size on each edge.
-    #[inline]
-    pub const fn splat(size: u16) -> Self {
-        Margins::hv_splat(size, size)
-    }
-
-    /// Margins via horizontal and vertical sizes
-    #[inline]
-    pub const fn hv(horiz: (u16, u16), vert: (u16, u16)) -> Self {
-        Margins { horiz, vert }
-    }
-
-    /// Margins via horizontal and vertical sizes
-    #[inline]
-    pub const fn hv_splat(h: u16, v: u16) -> Self {
-        Margins {
-            horiz: (h, h),
-            vert: (v, v),
-        }
-    }
-
-    /// Sum of horizontal margins
-    #[inline]
-    pub fn sum_horiz(&self) -> i32 {
-        i32::from(self.horiz.0) + i32::from(self.horiz.1)
-    }
-
-    /// Sum of vertical margins
-    #[inline]
-    pub fn sum_vert(&self) -> i32 {
-        i32::from(self.vert.0) + i32::from(self.vert.1)
-    }
-
-    /// Pad a size with margins
-    pub fn pad(self, size: Size) -> Size {
-        Size::new(size.0 + self.sum_horiz(), size.1 + self.sum_vert())
-    }
-}
-
-/// Policy for stretching widgets beyond ideal size
-#[derive(Copy, Clone, Debug, PartialEq, Eq, Ord, PartialOrd, Hash)]
-pub enum StretchPolicy {
-    /// Do not exceed ideal size
-    Fixed,
-    /// Can be stretched to fill space but without utility
-    Filler,
-    /// Extra space has low utility
-    LowUtility,
-    /// Extra space has high utility
-    HighUtility,
-    /// Greedily consume as much space as possible
-    Maximize,
-}
-
-impl Default for StretchPolicy {
-    fn default() -> Self {
-        StretchPolicy::Fixed
-    }
-}
 
 /// Widget sizing information
 ///
